% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{curr_meas}
\alias{curr_meas}
\title{Current Water Level Measurements from PegelOnline}
\format{
A tibble with 680 rows and 6 variables:
\describe{
  \item{station}{Name of the river station (character).}
  \item{uuid}{Unique station identifier (character).}
  \item{timestamp}{Date and time of the measurement (POSIXct / datetime).}
  \item{wl}{Water level in cm (numeric).}
  \item{stateMnwMhw}{Status relative to mean low/high water (character).}
  \item{stateNswHsw}{Status relative to normal / high water (character).}
}
}
\source{
\url{https://www.pegelonline.wsv.de/webservices/}
}
\usage{
curr_meas
}
\description{
This dataset contains water level measurements from German river stations, 
including timestamps, water level, and status indicators.
}
\examples{
library(dplyr)
data(curr_meas)
curr_meas \%>\% filter(station == "CELLE") \%>\% head(3)
}
\keyword{datasets}
