% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_path_manager.R
\name{first_installation}
\alias{first_installation}
\title{Initialize cache files on first installation}
\usage{
first_installation()
}
\value{
This function is called for its side effects. It does not return a value.
}
\description{
This function ensures that required cache files are available in the user's
cache directory for the \code{dispPO} package. If the cache directory does not
exist, it will be created. If specific cache files are missing, they will be
copied from the package's internal \code{inst/extdata/cache} directory.
}
\details{
The function checks for the presence of the following files in the user cache:
\itemize{
  \item \code{curr_meas.rds}
  \item \code{lst_query.rds}
  \item \code{wl_list.rds}
}
If any of these files are not found, they are copied from the installed
package directory.
}
\examples{
\dontrun{
first_installation()
}
}
